Moving networking code to FClient
Added FClient and removed Fuego. Made FStore only responsible for managing directory state and calling FClient’s networking code.
